+++ bomberman/src/level_tile_matrix_tools/coordinate_translator.c
@@ -1,12 +1,12 @@
 #include "../../headers/level_tile_matrix_tools.h"
 #include "../../headers/sprite_properties.h"
 
-int pixelFromTile(int t)
+float pixelFromTile( int t )
 {
-    return t * TILE_SIZE;
+    return (float) (t * TILE_SIZE);
 }
 
-int tileFromPixel(int p)
+int tileFromPixel( float p )
 {
-    return p / TILE_SIZE;
+    return (int) (p / TILE_SIZE);
 }
